Front End Registration Styling Adjustments
Adjustments to front end registration module to make adjustments for responsiveness.  Adjusted location and styling of Payment Information, push and pulled columns on eventure list, made adjustments to the navbar.  Will require additional work for functionality.  When the cart is clicked on a tablet or smartphone device it should take you to a cart page or possible confirm.  Discussion needed.