Working required flag
The required flag is working in all cases including radio buttons.  Radio
buttons are a bit of trouble, and I had to make a decision with how to handle them.
When a radio button and a check box are required, I default them to an answer. This
is because the default validator does not show their required state well.