Refactored file tree to put datacontext inside common
This refactoring is to allow the registraiton app to not include the dashboard app, but still get at data context.
This seperation of concerns will then allow the dashboard to include the registraiton for manual reg etc.
Without this, there would probably have been a race condition where each one tried to load eachother in
an infinite loop.