Spinner Works with Pages that Use it or Not
I rearranged the spinner code to work when you want it and when you don't.
This was accomplished by removing the code that made it start spinning when
the router changes pages.  Now, it only starts spinning when you execute
the `activateController` function.  Before `activateController` resolves the
promises, it starts the spinner.  Then after the promises finish, it hides
the spinner. This new procedure results in the spinner only showing when
you want to use it and not causing problems when you don't want to use it.