Adds role based authentication to Manual Registration
This applies role based authentication to the Manual Registration without applying this to the
regular registration process.  It provides this functionality by adding the role settings to
the registration routes, but not applying them in the registration.  Further, it seperates
route definition from route application, which allows for route addition only when a certial
asset is included (./registration/config.js).

Closes #125